requirements.txt
inspect-ai
matplotlib
scipy
pyyaml
